@@ -1,4 +1,5 @@
-//+build windows
+//go:build windows
+// +build windows
 
 package windows // import "github.com/docker/docker/daemon/graphdriver/windows"
 
