@@ -1,5 +1,5 @@
-// +build linux,cgo,!static_build
-// +build !libdm_dlsym_deferred_remove,!libdm_no_deferred_remove
+//go:build linux && cgo && !static_build && !libdm_dlsym_deferred_remove && !libdm_no_deferred_remove
+// +build linux,cgo,!static_build,!libdm_dlsym_deferred_remove,!libdm_no_deferred_remove
 
 package devicemapper // import "github.com/docker/docker/pkg/devicemapper"
 
