@@ -1,5 +1,5 @@
-// +build linux,cgo
-// +build !libdm_dlsym_deferred_remove,libdm_no_deferred_remove
+//go:build linux && cgo && !libdm_dlsym_deferred_remove && libdm_no_deferred_remove
+// +build linux,cgo,!libdm_dlsym_deferred_remove,libdm_no_deferred_remove
 
 package devicemapper // import "github.com/docker/docker/pkg/devicemapper"
 
